refactor: add usePasswordGenerator
